renamed ChatHandler --> ChatHandlerWebsocket
